- tcs の f32 数値モデル (T226) に追従し、stub / cs-lib / samples / .NET の
  host と facade を float / int にする。generator は C# 側も float を出し、
  facade の double 変換を消す。xorshift は int32 wrap で同じ列を保つ
- ci.yml に gate の追加分 (gen-api / gen-lubx の check、.NET digest と golden、
  ngs、raw Lua) を master の native-gate.sh の並列構造で組み込み、Haxe の段取り
  (setup-haxe5、haxe-wasm の build と cache、gen-haxe) を落とす。Windows は
  golden の前に dotnet を入れ、00_hello の .NET digest 比較を足す
- backend 名は d3d12 / vulkan / sdlgpu (native 廃止)。C API の config 検証、
  shader target の選択、サンプルの LUB_BACKEND 既定、README を合わせる
- tcs submodule を master 取り込み後の agent/lua-naming に向ける。playground の
  verify は master の shard 方式に C# だけの流れを載せ直す
- golden: tcs→Lua は不変。.NET 実行は float 化で tcs→Lua と byte 一致になった
  ので 7 枚を撮り直す

+++ cs-lib/lubx/Bones.cs
@@ -17,7 +17,7 @@ public static class Bones
 
     /// <summary>pivot (px, py, pz) 回りの回転 (model 空間)。
     /// T(p) · R · T(−p)。</summary>
-    public static Mat4 PivotRot(double px, double py, double pz, Mat4 rot)
+    public static Mat4 PivotRot(float px, float py, float pz, Mat4 rot)
     {
         return Mat4.Translate(new Vec3(px, py, pz))
             * rot * Mat4.Translate(new Vec3(-px, -py, -pz));
@@ -27,10 +27,10 @@ public static class Bones
     /// mat4 × 8 = 128 float に詰める。resolve が null を返した bone は
     /// 単位行列。(x, y, z) はその bone の pivot (pivotRot にそのまま
     /// 渡せる)。</summary>
-    public static List<double> Pack(MeshData? mesh,
-        Func<string, double, double, double, Mat4?> resolve)
+    public static List<float> Pack(MeshData? mesh,
+        Func<string, float, float, float, Mat4?> resolve)
     {
-        var arr = new List<double>();
+        var arr = new List<float>();
         int count = 0;
         if (mesh != null && mesh.Bones != null)
         {
