+++ cs-lib/lubx/MeshText.cs
@@ -1,7 +1,7 @@
 // 実装ライブラリ lubx の MeshText。
 // utf8 / string 関数は stub の utf8 / @string class で呼ぶ (@string は
 // using static で取り込み len(s) と裸で呼ぶ)。
-// gm.positions / gm.indices は List<double> を 0-based で書く (tcs の
+// gm.positions / gm.indices は List<float> を 0-based で書く (tcs の
 // indexer が +1 して Lua の 1-based に写る)。
 // char は C# 予約語で、@char は tcs が宣言をそのまま `function MeshText:@char`
 // と emit して不正 Lua になるため Char にしている。
@@ -15,9 +15,9 @@ public class GlyphEntry
     public BufferRef? Vb;
     public BufferRef? Ib;
     public int Count;
-    public double Advance;
-    public double Cx;
-    public double Cy;
+    public float Advance;
+    public float Cx;
+    public float Cy;
 }
 
 /// <summary>メッシュグリフ描画 (大サイズレジーム)。TTF 輪郭を三角形化して
@@ -105,22 +105,22 @@ public class MeshText
             {
                 Count = 0,
                 Advance = gm.Advance,
-                Cx = 0.0,
-                Cy = 0.0,
+                Cx = 0.0f,
+                Cy = 0.0f,
             };
             glyphs[cp] = empty;
             return empty;
         }
-        var verts = new List<double>();
-        double minX = 1e9;
-        double minY = 1e9;
-        double maxX = -1e9;
-        double maxY = -1e9;
+        var verts = new List<float>();
+        float minX = 1e9f;
+        float minY = 1e9f;
+        float maxX = -1e9f;
+        float maxY = -1e9f;
         for (int i = 0; i < gm.VertCount; i++)
         {
             // vertex i の x, y (stride 3、z は捨てる)
-            double x = gm.Positions[i * 3];
-            double y = gm.Positions[i * 3 + 1];
+            float x = gm.Positions[i * 3];
+            float y = gm.Positions[i * 3 + 1];
             verts.Add(x);
             verts.Add(y);
             if (x < minX)
@@ -132,8 +132,8 @@ public class MeshText
             if (y > maxY)
                 maxY = y;
         }
-        // use_buffer は List<double> を取るので indices を詰め替える
-        var idx = new List<double>();
+        // use_buffer は List<float> を取るので indices を詰め替える
+        var idx = new List<float>();
         for (int i = 0; i < gm.IndexCount; i++)
             idx.Add(gm.Indices[i]);
         var e = new GlyphEntry
@@ -142,8 +142,8 @@ public class MeshText
             Ib = Gfx.UseBuffer(key + "_i:" + cp, Gfx.BufferType.Index, idx, version),
             Count = gm.IndexCount,
             Advance = gm.Advance,
-            Cx = (minX + maxX) * 0.5,
-            Cy = (minY + maxY) * 0.5,
+            Cx = (minX + maxX) * 0.5f,
+            Cy = (minY + maxY) * 0.5f,
         };
         glyphs[cp] = e;
         return e;
@@ -151,14 +151,14 @@ public class MeshText
 
     private static Color ColorOrWhite(Color? c)
     {
-        return c ?? Color.Rgb(1.0, 1.0, 1.0);
+        return c ?? Color.Rgb(1.0f, 1.0f, 1.0f);
     }
 
     /// <summary>グリフ 1 つ。(x, y) は centered=false ならベースライン原点、
     /// true なら bbox 中心を (x, y) に置く。size は px/em、angle は CCW
     /// ラジアン。</summary>
-    public void Glyph(int cp, double x, double y, double size,
-        double? angle = null, Color? tint = null, bool? centered = null)
+    public void Glyph(int cp, float x, float y, float size,
+        float? angle = null, Color? tint = null, bool? centered = null)
     {
         var sh = Ensure();
         if (sh == null)
@@ -178,13 +178,13 @@ public class MeshText
             ["indices"] = ib,
             ["uniforms"] = new Dictionary<string, object>
             {
-                ["psr"] = new List<double> { x, y, size, angle ?? 0.0 },
-                ["tint"] = new List<double> { c.R, c.G, c.B, c.A },
-                ["screen"] = new List<double>
-                    { (double)logicalW, (double)logicalH, 0.0, 0.0 },
+                ["psr"] = new List<float> { x, y, size, angle ?? 0.0f },
+                ["tint"] = new List<float> { c.R, c.G, c.B, c.A },
+                ["screen"] = new List<float>
+                    { (float)logicalW, (float)logicalH, 0.0f, 0.0f },
                 ["center"] = ctr
-                    ? new List<double> { e.Cx, e.Cy, 0.0, 0.0 }
-                    : new List<double> { 0.0, 0.0, 0.0, 0.0 },
+                    ? new List<float> { e.Cx, e.Cy, 0.0f, 0.0f }
+                    : new List<float> { 0.0f, 0.0f, 0.0f, 0.0f },
             },
         }, new DrawOpts
         {
@@ -197,8 +197,8 @@ public class MeshText
 
     /// <summary>文字列の先頭グリフ 1 つを描く。glyph() の String 版
     /// (char は C# の予約語のため Char)。</summary>
-    public void Char(string s, double x, double y, double size,
-        double? angle = null, Color? tint = null, bool? centered = null)
+    public void Char(string s, float x, float y, float size,
+        float? angle = null, Color? tint = null, bool? centered = null)
     {
         foreach (var r in s.EnumerateRunes())
         {
@@ -208,7 +208,7 @@ public class MeshText
     }
 
     /// <summary>1 行をベースライン左端から。</summary>
-    public void Text(string s, double x, double baselineY, double size,
+    public void Text(string s, float x, float baselineY, float size,
         Color? tint = null)
     {
         var pen = x;
@@ -218,23 +218,23 @@ public class MeshText
             var e = GlyphFor(cp);
             if (e != null)
             {
-                Glyph(cp, pen, baselineY, size, 0.0, tint, false);
+                Glyph(cp, pen, baselineY, size, 0.0f, tint, false);
                 pen += e.Advance * size;
             }
         }
     }
 
     /// <summary>1 行を中央揃えで (cx は中心)。</summary>
-    public void TextCentered(string s, double cx, double baselineY, double size,
+    public void TextCentered(string s, float cx, float baselineY, float size,
         Color? tint = null)
     {
-        Text(s, cx - Width(s, size) * 0.5, baselineY, size, tint);
+        Text(s, cx - Width(s, size) * 0.5f, baselineY, size, tint);
     }
 
     /// <summary>1 行の幅 (px)。advance の合計 × size。</summary>
-    public double Width(string s, double size)
+    public float Width(string s, float size)
     {
-        var sum = 0.0;
+        var sum = 0.0f;
         foreach (var r in s.EnumerateRunes())
         {
             var e = GlyphFor(r.Value);
